requirements.txt
flask==2.2.2
flask-bcrypt==1.0.1
python-dotenv==0.20.0
requests==2.28.1
pymysql==1.0.2
pymongo
duffel-api==0.6.2
bcrypt==4.0.0
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.1
werkzeug==2.2.2
click==8.1.3
colorama==0.4.5
setuptools
wheel
pywhatkit